Updated README and restructured project
