+++ start-backend.py
@@ -1,0 +1,4 @@
+from backend import app 
+
+if __name__ == '__main__':
+    app.run()
